Created first split between humans and AI. Battle is having issues with ending too soon, but I want to finish the divide before fixing that.
