+++ src/main/java/util/CombatLog.java
@@ -1,5 +1,6 @@
 package util;
 
+import entities.AbstractPlayer;
 import static java.lang.System.out;
 import java.util.ArrayList;
 import entities.Projectile;
@@ -17,13 +18,14 @@ public class CombatLog{
 		}
 	}
 	public static void logProjectileData(Projectile p){
-		add("-------");
+		AbstractPlayer hit = p.getHit();
+        add("-------");
 		add("*Projectile:");
 		add("Generated by attack instance #" + p.getUseId());
 		add("Bearing " + p.getDir().getDegrees() + " degrees");
 		add("Generated from " + p.getAttackName());
 		add("by " + p.getUser().getName());
-		add("hitting " + p.getHit().getName());
+		add("hitting " + ((hit == null) ? "nobody" : p.getHit().getName()));
 		add("after traveling " + p.getDistance() + " pixels.");
 	}
 	public static void displayLog(){
